backend/requirements.txt
Flask>=2.2.5
Flask_Cors==3.0.10
Flask_Mail==0.9.1
flask_restx==1.0.6
flask_sqlalchemy==3.0.5
PyJWT==2.7.0
SQLAlchemy==2.0.4
python-dotenv>=0.19.1
flask_cors==3.0.10  # Add this line